Predicting Airline Delays with Hadoop
One of the main goals is using machine learning algorithms to build predictive models with Python packages and data analysis programs. Training the original datasets is impor- tant to build models with its performance. Finding a good combination of technologies and programming languages would be cruicial to make a successful project.
Dataset The data can be downloaded from Bureau of Transportation Statistics where it is described in detail. An other link to more detailed data can be found here.
Possible tools
- Apache Pig - Hadoop
- Python
- scikit-learn